Refresh compile_commands.json during the build (#1098)
CMake compile_commands.json file is not found even though it has been
created. During a CMake build's generation stage, it generates the
compile_commands.json file inside the build folder. However,
when it comes to be used it is not found because it hasn't been
reliably refreshed.

To solve this, the file resource is refreshed

Co-authored-by: John Moule <evil_bandit_betamax@yahoo.co.uk>