Add accessor for included pty instance (#1109)
Currently if I get passed a Process from by a method call I can check it
for the Spawner and can already call special methods like hangup(). But
there is no way to get access the the PTY used to create the Spawner so
for example one can't call PTY#setTerminalSize or query any other
properties.

This now adds a new method pty() that returns this instance for further
investigation and actions.