Fixes User Defined Toolchain clang type forgetfullnes
When a User Defined Toolchain of type clang was created (Preferences >
C/C++ > Core Build Toolchains > User Defined Toolchains) the Type was
reported correctly as clang. But after restarting the workbench the Type
was reported as GCC. This is now fixed so the Type always shows as clang
for a clang toolchain.