Add Terminal's native fragment for Linux riscv64
- add a new fragment 'org.eclipse.cdt.core.linux.riscv64' containing
compiled riscv64 natives (not included) suitable for the Linux riscv64
platform.

- add the 'libserial.so' binary (not included) compiled for
Linux riscv64, for the serial communication interface.

- add the linux/gtk/riscv64 environment triplet to various Maven
build scripts to support building CDT software for that environment.

Note that the various *.so binary files are not included in this commit
and will be provided in subsequent commits with the corresponding
update to CDT build infrastructure.

Fixes #980