Do not require org.eclipse.tm.terminal.view.core in the feature (#1203)
The plugin org.eclipse.tm.terminal.view.core is moved to platform and
should therefore be handled like any other third party dependency. It
actually is already part of the feature imports and the
org.eclipse.tm.terminal.view.ui require bundle it, so mentioning that
one is actually redundant and fixes it to that particular version
otherwhise.