Gdb Manual Remote launch targets for CoreBuild projects
Support the Gdb Remote TCP and Serial launch targets for CoreBuild
projects.  A new non-public LaunchConfiguration type is added, that is
working with the same launch descriptor type as CoreBuild Local
projects. Also a new LaunchConfigurationProvider that supports the Gdb
Remote TCP and Serial targets is added.

It makes use of the launch delegate
org.eclipse.cdt.dsf.gdb.launching.GdbTargetedLaunchDelegate, which is
also used for standard GDB (DSF) Manual Remote Debugging launch
configurations for TCP and Serial connections.

For the user is appears as if there is a single CoreBuild launch
configuration that adapts dynamically to the launch target, while in
reality there is a separate launch configuration for each target.

Fixes #1168